fix: share CojView with version in state
