+++ components/main/Main.tsx
@@ -24,8 +24,7 @@ import Programmes from "../programmes/Programmes";
 import Breadcrumbs from "../breadcrumbs/Breadcrumbs";
 import Dsp from "../dsp/Dsp";
 import GlobalAlert from "./GlobalAlert";
-import CojView9 from "../forms/conditionOfJoining/CojView9";
-import CojView10 from "../forms/conditionOfJoining/CojView10";
+import CojView from "../forms/conditionOfJoining/CojView";
 import TSSHeader from "../navigation/TSSHeader";
 import packageJson from "../../package.json";
 import { loadFormAList } from "../../redux/slices/formASlice";
